Quote variables exported into ENV
- This allows support for multi-line secrets like SSL certificates and
the like. The idea being that you can scrape full file contents into
memory for encryption to be reproduceible in its entirety, without
pre-base64 encoding and encrypting the base64 encoded secret.

This makes it "less fiddly" when we can use things like helm to base64
encode the values for us anyway.